Improved archery minigame reward system (described below).
The archery minigame in its previous iteration inherently gave the first
player (and others, if there are more than 2 players) a disadvantage.
Imagine you had two players who both made it to wave 10. Player 1 shoots
first. If he misses, player 2 automatically wins without having to
actually shoot the target. Imagine that player 2 also would have missed.
If the player order were reversed, then player 2 would have won, even
though neither of them could make the shot. To counteract this, the
reward system has been reworked as described in the next paragraph.

Now, instead of ending as soon as there is only one player standing, the
game ends the first time one of the following occurs:
 - a round starts where there is only 1 player standing
 - there are 0 players standing
If multiple players miss on the same wave, they will all receive the
same medal. Therefore, players can be "grouped" by the wave they made it
to, and every player in each group is given the same medal. This implies
that if there is 1 person standing but the round is not over, that
person must hit the target in order to receive a higher medal than the
other player(s) who missed on that round. Note that medals start at gold
and work down (this implies, for example, that if all 4 players miss on
the same wave, they will all receive golds).